pyproject.toml
[tool.poetry]
name = "message-summarizer"
version = "0.1.0"
description = ""
authors = ["Leroy <leroy.souz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.37.1"
python-dotenv = "^1.0.1"
discord = "^2.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"